+++ apps/drawing_canvas.hh
@@ -1,0 +1,12 @@
+// header file for drawing_canvas.hh
+#ifndef DRAWING_CANVAS_FUNCS
+#define DRAWING_CANVAS_FUNCS
+
+#include "imgui.h" // makefile include flag locates imgui so no need for relative path
+
+#include <vector>
+#include <set>
+
+void showDrawingCanvas();
+
+#endif
