+++ apps/image_editor.cpp
@@ -2,7 +2,7 @@
 #include "image_editor.hh"
 
 #define STB_IMAGE_IMPLEMENTATION
-#include "../external/stb_image.h"
+#include "stb_image.h"
 
 std::string trimString(std::string str){
 	std::string trimmed("");
