+++ artstation.cpp
@@ -2,23 +2,16 @@
 #include "imgui_impl_sdl.h"
 #include "imgui_impl_opengl3.h"
 
-#include "apps/image_editor.hh"
+#include "drawing_canvas.hh"
+#include "image_editor.hh"
+#include "3d_viewer.hh"
 
-#define TINYOBJLOADER_IMPLEMENTATION
-#include "external/tiny_obj_loader.h"
-
-#include <iostream>
 #include <chrono>
+#include <iostream>
 #include <vector>
-#include <set>
-#include <math.h>
 #include <stdio.h>
 #include <SDL.h>
 #include <GL/glew.h>
-#include <GLM/vec2.hpp>
-#include <GLM/vec3.hpp>
-#include <GLM/mat4x4.hpp>
-#include <GLM/ext.hpp>
 
 #if defined(IMGUI_IMPL_OPENGL_ES2)
 #include <SDL_opengles2.h>
@@ -26,548 +19,6 @@
 #include <SDL_opengl.h>
 #endif
 
-void showDrawingCanvas(){
-	// keep track of canvas data
-	static std::vector<ImVec2> canvasPoints;
-	
-	// use this to keep track of where each separately drawn segement ends because everything gets redrawn each re-render from the beginning
-	static std::set<int> lastSegmentIndexes;
-	
-	// keep track of colors
-	static std::vector<ImU32> canvasPointColors;
-	
-	// color wheel stuff
-	static ImVec4 color = ImVec4(1.0f, 1.0f, 1.0f, 1.0f); // note this isn't assignment but just initialization
-	static bool ref_color = false;
-	static ImVec4 ref_color_v(1.0f, 0.0f, 1.0f, 0.5f);
-	ImGuiColorEditFlags flags = ImGuiColorEditFlags_Float | ImGuiColorEditFlags_HDR;
-	ImU32 selectedColor = ImColor(color);
-	
-	//static ImVec2 scrolling(0.0f, 0.0f);
-	
-	ImGui::BeginChild("drawing canvas", ImVec2(0, 0), true);
-	ImGui::Columns(2, "stuff");
-	
-	ImVec2 canvas_pos0 = ImGui::GetCursorScreenPos();      // ImDrawList API uses screen coordinates!
-	ImVec2 canvas_size = ImGui::GetContentRegionAvail();   // Resize canvas to what's available
-	ImVec2 canvas_pos1 = ImVec2(canvas_size.x*0.95f, canvas_size.y*0.95f);
-
-	// Draw border and background color
-	ImGuiIO& io = ImGui::GetIO();
-	ImDrawList* draw_list = ImGui::GetWindowDrawList();
-	draw_list->AddRectFilled(canvas_pos0, canvas_pos1, IM_COL32(50, 50, 50, 255));
-	draw_list->AddRect(canvas_pos0, canvas_pos1, IM_COL32(255, 255, 255, 255));
-
-	// This will catch our interactions
-	ImGui::InvisibleButton("drawingCanvas", canvas_pos1, ImGuiButtonFlags_MouseButtonLeft | ImGuiButtonFlags_MouseButtonRight);
-	
-	// Hovered - this is so we ensure that we take into account only mouse interactions that occur
-	// on this particular canvas. otherwise it could pick up mouse clicks that occur on other windows as well.
-	const bool is_hovered = ImGui::IsItemHovered();
-	const bool is_active = ImGui::IsItemActive();   // Held
-	
-	const ImVec2 origin(canvas_pos0.x, canvas_pos0.y); // Lock scrolled origin
-	const ImVec2 mouse_pos_in_canvas(io.MousePos.x - origin.x, io.MousePos.y - origin.y);
-
-	if (is_active && is_hovered && ImGui::IsMouseClicked(ImGuiMouseButton_Left))
-	{
-		canvasPoints.push_back(mouse_pos_in_canvas);
-		canvasPointColors.push_back(selectedColor);
-	}
-	
-	ImVec2 drag_delta = ImGui::GetMouseDragDelta(ImGuiMouseButton_Left);
-	if (is_hovered && is_active && (drag_delta.x > 0 || drag_delta.y > 0))
-	{
-		canvasPoints.push_back(mouse_pos_in_canvas);
-		canvasPointColors.push_back(selectedColor);
-	}
-	
-	int lastIdx = -1;
-	if (is_hovered && ImGui::IsMouseReleased(ImGuiMouseButton_Left))
-	{
-		// I think we might be setting the last element in the canvasPoints vector to
-		// mouse_pos_in_canvas? and since it's a reference, it works? don't think we need this though
-		//canvasPoints.back() = mouse_pos_in_canvas;
-		lastIdx = canvasPoints.size() - 1;
-	}
-	
-	// need to redraw each event update
-	int numPoints = (int)canvasPoints.size();
-	for(int i = 0; i < numPoints; i++){
-		ImVec2 p1 = ImVec2(origin.x + canvasPoints[i].x, origin.y + canvasPoints[i].y);
-		ImVec2 p2 = ImVec2(origin.x + canvasPoints[i+1].x, origin.y + canvasPoints[i+1].y);
-		ImU32 theColor = canvasPointColors[i];
-		
-		if(lastSegmentIndexes.find(i) != lastSegmentIndexes.end()){
-			//std::cout << "last point index: " << lastPointIndex << std::endl;				
-			// don't draw a line from this point to the next
-			draw_list->AddCircleFilled(p1, 4.0f, theColor, 8.0f);
-			continue;
-		}
-		if(i < numPoints - 1){
-			// connect the points
-			draw_list->AddLine(p1, p2, theColor, 8.0f);
-		}else{
-			// just draw point
-			draw_list->AddCircle(p1, 4.0f, theColor, 8.0f);
-		}
-	}
-	
-	if(lastIdx > -1){
-		// keep track of the last point of the last segment drawn so we don't connect two separately drawn segments
-		lastSegmentIndexes.insert(lastIdx); //lastPointIndex = lastIdx;
-	}
-	
-	ImGui::NextColumn();
-	
-	// show color wheel
-	ImGui::ColorPicker4("MyColor", (float*)&color, flags, ref_color ? &ref_color_v.x : NULL);
-	
-	ImGui::Dummy(ImVec2(0.0f, 5.0f)); // add some vertical spacing
-	
-	if(ImGui::Button("clear")){
-		canvasPoints.clear();
-		lastSegmentIndexes.clear();
-		canvasPointColors.clear();
-	}
-	
-	ImGui::Columns();
-	
-	ImGui::EndChild();
-}
-
-
-// https://github.com/tinyobjloader/tinyobjloader
-void getObjModelInfo(
-	auto& shapes, 
-	auto& attrib, 
-	std::vector<glm::vec3>& vertices,
-	std::vector<glm::vec2>& uvs,
-	std::vector<glm::vec3>& normals
-	){
-	for (size_t s = 0; s < shapes.size(); s++) {
-	  // Loop over faces(polygon)
-	  size_t index_offset = 0;
-	  for (size_t f = 0; f < shapes[s].mesh.num_face_vertices.size(); f++) {
-		size_t fv = size_t(shapes[s].mesh.num_face_vertices[f]);
-
-		// Loop over vertices in the face.
-		for (size_t v = 0; v < fv; v++) {
-		  // access to vertex
-		  tinyobj::index_t idx = shapes[s].mesh.indices[index_offset + v];
-		  float vx = (float)attrib.vertices[3*size_t(idx.vertex_index)+0];
-		  float vy = (float)attrib.vertices[3*size_t(idx.vertex_index)+1];
-		  float vz = (float)attrib.vertices[3*size_t(idx.vertex_index)+2];
-		  vertices.push_back(glm::vec3(vx, vy, vz));
-
-		  // Check if `normal_index` is zero or positive. negative = no normal data
-		  if (idx.normal_index >= 0) {
-			float nx = (float)attrib.normals[3*size_t(idx.normal_index)+0];
-			float ny = (float)attrib.normals[3*size_t(idx.normal_index)+1];
-			float nz = (float)attrib.normals[3*size_t(idx.normal_index)+2];
-			normals.push_back(glm::vec3(nx, ny, nz));
-		  }
-
-		  // Check if `texcoord_index` is zero or positive. negative = no texcoord data
-		  if (idx.texcoord_index >= 0) {
-			float tx = (float)attrib.texcoords[2*size_t(idx.texcoord_index)+0];
-			float ty = (float)attrib.texcoords[2*size_t(idx.texcoord_index)+1];
-			uvs.push_back(glm::vec2(tx, ty));
-		  }
-
-		  // Optional: vertex colors
-		  // tinyobj::real_t red   = attrib.colors[3*size_t(idx.vertex_index)+0];
-		  // tinyobj::real_t green = attrib.colors[3*size_t(idx.vertex_index)+1];
-		  // tinyobj::real_t blue  = attrib.colors[3*size_t(idx.vertex_index)+2];
-		}
-		index_offset += fv;
-
-		// per-face material
-		//shapes[s].mesh.material_ids[f];
-	  }
-	}
-}
-
-// 3d stuff - TODO: move elsewhere
-// shaders
-const char* vertexShaderSource = R"glsl(
-	#version 330 core
-
-	layout(location = 0) in vec3 position;
-	layout(location = 1) in vec2 vertexUV;
-	
-	out vec2 uv;
-	out vec3 pos;
-	
-	uniform mat4 mvp;
-
-	void main()
-	{
-		pos = position;
-		uv = vertexUV;
-		gl_Position = mvp * vec4(position, 1.0);
-	}
-)glsl";
-
-const char* fragShaderSource = R"glsl(
-	#version 330 core
-
-	uniform float time;
-
-	in vec3 pos;
-	out vec4 outColor;
-
-	void main()
-	{
-		float newR = cos(time)/2; //cos(time*pos.x);
-		float newG = sin(time)/2; //sin(time*pos.y);
-		float newB = cos(time)/2; //cos(time*pos.y);
-		outColor = vec4(newR, newG, newB, 1.0);
-	}
-)glsl";
-
-//GLuint shaderProgram;
-void setupShaders(GLuint& shaderProgram){
-	// compile vertex shader
-	GLuint vertexShader = glCreateShader(GL_VERTEX_SHADER);
-	glShaderSource(vertexShader, 1, &vertexShaderSource, NULL);
-	glCompileShader(vertexShader);
-	
-	GLint status;
-	
-	glGetShaderiv(vertexShader, GL_COMPILE_STATUS, &status);
-	if(status != GL_TRUE){
-		std::cout << "error compiling vertex shader!" << std::endl;
-	}
-	
-	// compile frag shader
-	GLuint fragmentShader = glCreateShader(GL_FRAGMENT_SHADER);
-	glShaderSource(fragmentShader, 1, &fragShaderSource, NULL);
-	glCompileShader(fragmentShader);
-	
-	glGetShaderiv(fragmentShader, GL_COMPILE_STATUS, &status);
-	if(status != GL_TRUE){
-		std::cout << "error compiling fragment shader!" << std::endl;
-	}
-	
-	// combine shaders into a program
-	shaderProgram = glCreateProgram();
-	glAttachShader(shaderProgram, vertexShader);
-	glAttachShader(shaderProgram, fragmentShader);
-	
-	glBindFragDataLocation(shaderProgram, 0, "outColor");
-	glLinkProgram(shaderProgram);
-	glUseProgram(shaderProgram);
-	
-	// set up position attribute in vertex shader
-	GLint posAttrib = glGetAttribLocation(shaderProgram, "position");
-	glVertexAttribPointer(posAttrib, 3, GL_FLOAT, GL_FALSE, 0, 0);
-	glEnableVertexAttribArray(posAttrib);
-}
-
-void setupCube(GLuint& vbo, GLuint& vao){
-	float vertices[] = {
-		-1.0f,-1.0f,-1.0f, // triangle 1 : begin
-		-1.0f,-1.0f, 1.0f,
-		-1.0f, 1.0f, 1.0f, // triangle 1 : end
-		1.0f, 1.0f,-1.0f, // triangle 2 : begin
-		-1.0f,-1.0f,-1.0f,
-		-1.0f, 1.0f,-1.0f, // triangle 2 : end
-		1.0f,-1.0f, 1.0f,
-		-1.0f,-1.0f,-1.0f,
-		1.0f,-1.0f,-1.0f,
-		1.0f, 1.0f,-1.0f,
-		1.0f,-1.0f,-1.0f,
-		-1.0f,-1.0f,-1.0f,
-		-1.0f,-1.0f,-1.0f,
-		-1.0f, 1.0f, 1.0f,
-		-1.0f, 1.0f,-1.0f,
-		1.0f,-1.0f, 1.0f,
-		-1.0f,-1.0f, 1.0f,
-		-1.0f,-1.0f,-1.0f,
-		-1.0f, 1.0f, 1.0f,
-		-1.0f,-1.0f, 1.0f,
-		1.0f,-1.0f, 1.0f,
-		1.0f, 1.0f, 1.0f,
-		1.0f,-1.0f,-1.0f,
-		1.0f, 1.0f,-1.0f,
-		1.0f,-1.0f,-1.0f,
-		1.0f, 1.0f, 1.0f,
-		1.0f,-1.0f, 1.0f,
-		1.0f, 1.0f, 1.0f,
-		1.0f, 1.0f,-1.0f,
-		-1.0f, 1.0f,-1.0f,
-		1.0f, 1.0f, 1.0f,
-		-1.0f, 1.0f,-1.0f,
-		-1.0f, 1.0f, 1.0f,
-		1.0f, 1.0f, 1.0f,
-		-1.0f, 1.0f, 1.0f,
-		1.0f,-1.0f, 1.0f
-	};
-	
-	// vertex buffer object
-	glBindBuffer(GL_ARRAY_BUFFER, vbo);
-	glBufferData(GL_ARRAY_BUFFER, sizeof(vertices), vertices, GL_STATIC_DRAW);
-	glBindBuffer(GL_ARRAY_BUFFER, 0); // unbind
-}
-
-
-void setupTriangle(GLuint& vbo, GLuint& vao){
-	float vertices[] = {
-		0.0f, 0.5f, 0.0f,
-		0.5f, 0.0f, 0.0f,
-		-0.5f, 0.0f, 0.0f
-	};
-	
-	// vertex buffer object
-	glBindBuffer(GL_ARRAY_BUFFER, vbo);
-	glBufferData(GL_ARRAY_BUFFER, sizeof(vertices), vertices, GL_STATIC_DRAW);
-	glBindBuffer(GL_ARRAY_BUFFER, 0); // unbind
-}
-
-// return the GLuint of the texture used for offscreen rendering
-// TODO: pass in dimensions later?
-void setupOffscreenFramebuffer(GLuint* frameBuffer, GLuint* texture){
-	// render scene to frame buffer -> texture -> use texture as image to render in the GUI
-	// create and bind frame buffer
-	glGenFramebuffers(1, frameBuffer);
-	glBindFramebuffer(GL_FRAMEBUFFER, *frameBuffer);
-	
-	// create the texture that the scene will be rendered to (it will be empty initially)
-	glActiveTexture(GL_TEXTURE1);
-	glGenTextures(1, texture);
-	glBindTexture(GL_TEXTURE_2D, *texture); // any gl texture operations now will use this texture
-	glTexImage2D(GL_TEXTURE_2D, 0, GL_RGB, 500, 500, 0, GL_RGB, GL_UNSIGNED_BYTE, NULL);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MIN_FILTER, GL_LINEAR);
-	glTexParameteri(GL_TEXTURE_2D, GL_TEXTURE_MAG_FILTER, GL_LINEAR);  
-	glBindTexture(GL_TEXTURE_2D, 0); // unbind the texture
-
-	// configure frame buffer to use texture
-	glFramebufferTexture2D(GL_FRAMEBUFFER, GL_COLOR_ATTACHMENT0, GL_TEXTURE_2D, *texture, 0);
-	
-	// depth buffer
-	GLuint depth;
-	glGenRenderbuffers(1, &depth);
-	glBindRenderbuffer(GL_RENDERBUFFER, depth);
-	glRenderbufferStorage(GL_RENDERBUFFER, GL_DEPTH_COMPONENT, 500, 500);
-	glBindRenderbuffer(GL_RENDERBUFFER, 0); // unbind after allocating storage
-	glFramebufferRenderbuffer(GL_FRAMEBUFFER, GL_DEPTH_ATTACHMENT, GL_RENDERBUFFER, depth);
-	
-	// unbind framebuffer now since we don't need it yet (we just needed to bind it to set it up)
-	glBindFramebuffer(GL_FRAMEBUFFER, 0);
-	
-	//GLenum drawBuffers[1] = {GL_COLOR_ATTACHMENT0};
-	//glDrawBuffers(1, drawBuffers);
-}
-
-float angleToRads(float angleInDeg){
-	float pi = 3.14159;
-	return (pi / 180) * angleInDeg;
-}
-
-glm::mat4 createPerspectiveMatrix(float fovAngle, float aspect, float near, float far){
-	glm::mat4 m(0.0);
-	
-	float fovRadians = angleToRads(fovAngle);
-	float f = 1.0f / tan(fovRadians/2.0f);
-	
-	m[0][0] = f / aspect;
-	m[1][1] = f;
-	m[2][2] = (far + near)/(near - far);
-	m[2][3] = -1.0f;
-	m[3][2] = (2.0f * far * near) / (near - far);
-
-	return m;
-}
-
-void show3dModelViewer(
-	GLuint offscreenFrameBuf, 
-	GLuint offscreenTexture, 
-	GLuint shaderProgram,
-	GLuint vbo,
-	GLuint vao,
-	GLuint uvBuffer,
-	GLuint matrixId,
-	auto startTime
-	){
-		
-	ImGui::BeginChild("3d model viewer", ImVec2(0, 800), true);
-	
-	static bool toggleWireframe = false;
-	
-	std::string filepath = "assets/battleship.obj";
-	tinyobj::ObjReaderConfig config;
-	config.mtl_search_path = "./";
-	
-	tinyobj::ObjReader reader;
-	
-	if(!reader.ParseFromFile(filepath, config)){
-		if(!reader.Error().empty()){
-			ImGui::Text("TinyObjReader: %s", (reader.Error()).c_str());
-			ImGui::EndChild();
-		}
-	}else{
-		if(!reader.Warning().empty()){
-			ImGui::Text("TinyObjReader: %s", (reader.Warning()).c_str());
-		}
-		
-		// set up some static variables to let user control camera
-		static float cameraCurrX = 0.0f;
-		static float cameraCurrY = 0.0f;
-		static float cameraCurrZ = -15.0f;
-		
-		auto& attrib = reader.GetAttrib();
-		auto& shapes = reader.GetShapes();
-		//auto& materials = reader.GetMaterials();
-		std::vector<glm::vec3> verts;
-		std::vector<glm::vec2> uvs;
-		std::vector<glm::vec3> normals;
-		
-		ImGui::Text("model has %d shapes", shapes.size());
-		ImGui::Text("model has %d vertices", attrib.vertices.size());
-		
-		getObjModelInfo(shapes, attrib, verts, uvs, normals);
-		
-		// uvs
-		//glBindBuffer(GL_ARRAY_BUFFER, uvBuffer);
-		//glBufferData(GL_ARRAY_BUFFER, sizeof(uvs)*sizeof(glm::vec2), &uvs[0], GL_STATIC_DRAW);
-		
-		if(glCheckFramebufferStatus(GL_FRAMEBUFFER) != GL_FRAMEBUFFER_COMPLETE){
-			std::cout << "framebuffer error: " << glCheckFramebufferStatus(GL_FRAMEBUFFER) << std::endl;
-		}else{
-			// draw to the offscreen frame buffer
-			// adjust the glviewport to be drawn to so the image comes out correctly
-			glBindFramebuffer(GL_FRAMEBUFFER, offscreenFrameBuf);
-			glViewport(0, 0, 500, 500);
-			glClearColor(0.62f, 0.73f, 0.78f, 0.0f);
-			glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
-			
-			if(toggleWireframe){
-				glPolygonMode(GL_FRONT_AND_BACK, GL_LINE);
-			}else{
-				glPolygonMode(GL_FRONT_AND_BACK, GL_FILL);
-			}
-			
-			glEnable(GL_DEPTH_TEST);
-			glDepthFunc(GL_LESS);
-			glEnable(GL_CULL_FACE);
-			
-			// load the vertices for the model
-			glBindBuffer(GL_ARRAY_BUFFER, vbo);
-			glBufferData(GL_ARRAY_BUFFER, verts.size()*sizeof(glm::vec3), &verts[0], GL_STATIC_DRAW);
-			
-			// for shaders
-			auto now = std::chrono::high_resolution_clock::now();
-			float time = std::chrono::duration_cast<std::chrono::duration<float>>(now - startTime).count();
-			GLuint t = glGetUniformLocation(shaderProgram, "time");
-			glUniform1f(t, time);
-			
-			// set up project matrix
-			float fovAngle = 60.0f;
-			float aspect = 1.0f; // 500 x 500 screen
-			float near = 0.01f;
-			float far = 100.0f;
-			glm::mat4 perspectiveMat = createPerspectiveMatrix(fovAngle, aspect, near, far);
-			
-			// set up modelview matrix (and do all the necessary transformations)
-			// store y rotation and rotate based on last y rotation so the model will rotate 360
-			static glm::mat4 lastYRot = glm::mat4(1.0);
-			lastYRot = glm::rotate(lastYRot, angleToRads(0.5f), glm::vec3(0,1,0));
-			
-			// the following should do: move model to origin, scale it, rotate about x, rotate about y, then move it to final position (i.e. further away from the camera along z)
-			// these operations here are ones we want to keep constant through each render, so we do them on the identity matrix (not using the previous object's transformation matrix)
-			glm::mat4 mvp = glm::mat4(1.0);
-			mvp = glm::translate(mvp, glm::vec3(cameraCurrX, cameraCurrY, cameraCurrZ));
-			mvp = mvp * lastYRot; // rotate about Y based on last rotation
-			mvp = glm::rotate(mvp, angleToRads(180.0f), glm::vec3(1, 0, 0)); // rotate about x 180 deg to make model right side up
-			mvp = glm::scale(mvp, glm::vec3(0.5f, 0.5f, 0.5f)); // scale down by half
-			mvp = glm::translate(mvp, glm::vec3(0, 0, 0)); // place at center (0,0,0)
-			
-			// multiply perspectiveMat and modelview to get the final view
-			perspectiveMat = perspectiveMat * mvp;
-			
-			glUniformMatrix4fv(matrixId, 1, GL_FALSE, &perspectiveMat[0][0]);
-			
-			// set up attributes to vertex buffer
-			glEnableVertexAttribArray(0);
-			glBindBuffer(GL_ARRAY_BUFFER, vbo);
-			glVertexAttribPointer(0, 3, GL_FLOAT, GL_FALSE, 0, (void*)0);
-			
- 			//glEnableVertexAttribArray(1);
-			//glBindBuffer(GL_ARRAY_BUFFER, uvBuffer);
-			//glVertexAttribPointer(1, 2, GL_FLOAT, GL_FALSE, 0, (void*)0);
-			
-			glDrawArrays(GL_TRIANGLES, 0, verts.size());
-			
-			glDisableVertexAttribArray(0);
-			//glDisableVertexAttribArray(1);
-			
-			// grab the offscreen texture and draw it to an imgui image
-			glActiveTexture(GL_TEXTURE1);
-			int w = 500;
-			int h = 500;
-			int pixelDataLen = w*h*3;
-			unsigned char* pixelData = new unsigned char[pixelDataLen];
-			glGetTexImage(GL_TEXTURE_2D, 0, GL_RGB, GL_UNSIGNED_BYTE, pixelData);
-			
-			// TODO: center the image?
-			//ImGui::SetCursorPos(ImVec2(ImGui::GetWindowSize().x/3, 0));
-			ImGui::Image((void *)(intptr_t)offscreenTexture, ImVec2(w, h));
-			
-			delete pixelData;
-			
-			// unbind offscreenFrameBuf and use default framebuffer again (show the gui window) - I think that's how this works?
-			glBindFramebuffer(GL_FRAMEBUFFER, 0);
-		}
-		
-		if(ImGui::Button("toggle wireframe")){
-			toggleWireframe = !toggleWireframe; 
-		}
-		
-		ImGui::Dummy(ImVec2(0.0f, 5.0f)); // add some vertical spacing
-		
-		// control camera x movement
-		ImGui::Text("move along X: ");
-		ImGui::SameLine();
-		if(ImGui::Button("+x")){
-			cameraCurrX += 0.5f;
-		}
-		ImGui::SameLine();
-		if(ImGui::Button("-x")){
-			cameraCurrX -= 0.5f;
-		}
-		ImGui::SameLine();
-		ImGui::Text("%f", cameraCurrX);
-		
-		// control camera y movement
-		ImGui::Text("move along Y: ");
-		ImGui::SameLine();
-		if(ImGui::Button("+y")){
-			cameraCurrY += 0.5f;
-		}
-		ImGui::SameLine();
-		if(ImGui::Button("-y")){
-			cameraCurrY -= 0.5f;
-		}
-		ImGui::SameLine();
-		ImGui::Text("%f", cameraCurrY);
-		
-		// control camera z movement
-		ImGui::Text("move along Z: ");
-		ImGui::SameLine();
-		if(ImGui::Button("+z")){
-			cameraCurrZ += 0.5f;
-		}
-		ImGui::SameLine();
-		if(ImGui::Button("-z")){
-			cameraCurrZ -= 0.5f;
-		}
-		ImGui::SameLine();
-		ImGui::Text("%f", cameraCurrZ);
-	}
-	
-	ImGui::EndChild();
-}
 
 // Main code
 int main(int, char**)
@@ -672,9 +123,6 @@ int main(int, char**)
 	GLuint uvBuffer;
 	glGenBuffers(1, &uvBuffer);
 	
-	//setupCube(vbo, vao);
-	//setupTriangle(vbo, vao);
-	
 	GLuint shaderProgram;
 	setupShaders(shaderProgram);
 	
@@ -688,7 +136,7 @@ int main(int, char**)
 	setupOffscreenFramebuffer(&offscreenFrameBuf, &offscreenTexture);
 	
 	// for fun shaders
-	auto startTime = std::chrono::high_resolution_clock::now();
+	std::chrono::time_point<std::chrono::high_resolution_clock> startTime = std::chrono::high_resolution_clock::now();
 	
     // Main loop
     bool done = false;
